Fix of alarm page: Clear of Memos did not work anymore

+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/hibernate/AlarmRepositoryHibernate.java
@@ -287,7 +287,7 @@ public class AlarmRepositoryHibernate implements AlarmRepository, InitializingBe
     @Transactional
     public void removeStickyMemo(Integer alarmId) {
         OnmsAlarm onmsAlarm = m_alarmDao.get(alarmId);
-        if (onmsAlarm != null) {
+        if (onmsAlarm != null && onmsAlarm.getStickyMemo() != null) {
             m_memoDao.delete(onmsAlarm.getStickyMemo());
             onmsAlarm.setStickyMemo(null);
         }
@@ -300,7 +300,7 @@ public class AlarmRepositoryHibernate implements AlarmRepository, InitializingBe
     @Transactional
     public void removeReductionKeyMemo(int alarmId) {
         OnmsAlarm onmsAlarm = m_alarmDao.get(alarmId);
-        if (onmsAlarm != null) {
+        if (onmsAlarm != null && onmsAlarm.getReductionKeyMemo() != null) {
             m_memoDao.delete(onmsAlarm.getReductionKeyMemo());
             onmsAlarm.setReductionKeyMemo(null);
         }
